pyproject.toml
[tool.poetry]
name = "convergent-routes"
version = "0.1.0"
description = "a multi-step synthesis planning framework to enchance retrosynthetic planning for compound libraries, focusing on convergent routes, where multiple compounds share common int"
authors = ["Paula Torren-Peraire"]
readme = "README.md"
packages = [{include = "convergent_routes"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
networkx = "^3.1"
registry-factory = "0.1.2"
torch = "^1.8.1"
pytorch-lightning = "1.8.0"
ipython = "^8.14.0"
hydra-core = "^1.3.2"
pandas = ">=1.4.1, <=2.0.2"
rdkit = "~2022.09.1"
matplotlib = "^3.9.2"
pysmilesutils = {git = "https://github.com/MolecularAI/pysmilesutils.git"}
numpy = "~1.24.3"
scipy = "~1.10.1"
ssbenchmark = {git = "https://github.com/PTorrenPeraire/modelsmatter_modelzoo.git", rev = "update_dependencies"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
